+++ src/main/java/com/example/concertreservationsystem/domain/service/event/ConcertEventService.java
@@ -4,7 +4,7 @@ package com.example.concertreservationsystem.domain.service.event;
 import com.example.concertreservationsystem.domain.model.Concert;
 import com.example.concertreservationsystem.domain.model.ConcertEvent;
 import com.example.concertreservationsystem.domain.repo.ConcertRepository;
-import com.example.concertreservationsystem.domain.service.concert.SeatInitializer;
+import com.example.concertreservationsystem.domain.service.init.seat.SeatInitializer;
 import com.example.concertreservationsystem.infrastructure.persistence.JpaConcertEventRepository;
 import com.example.concertreservationsystem.application.event.dto.request.EventRequestDto;
 import com.example.concertreservationsystem.application.event.dto.response.EventResponseDto;
