+++ src/main/java/com/example/concertreservationsystem/domain/service/init/concert/InitDataConcertAndSeatService.java
@@ -3,9 +3,8 @@ package com.example.concertreservationsystem.domain.service.init.concert;
 import com.example.concertreservationsystem.domain.model.Concert;
 import com.example.concertreservationsystem.domain.model.ConcertEvent;
 import com.example.concertreservationsystem.domain.repo.ConcertRepository;
-import com.example.concertreservationsystem.domain.service.concert.SeatInitializer;
+import com.example.concertreservationsystem.domain.service.init.seat.SeatInitializer;
 import com.example.concertreservationsystem.infrastructure.persistence.JpaConcertEventRepository;
-import com.example.concertreservationsystem.infrastructure.persistence.JpaConcertRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
